feat: Implement Prospect Tag management functionality
