+++ app-modules/prospect/src/Filament/Resources/ProspectTags/Pages/CreateProspectTag.php
@@ -34,53 +34,34 @@
 </COPYRIGHT>
 */
 
-namespace AdvisingApp\Prospect\Filament\Resources;
+namespace AdvisingApp\Prospect\Filament\Resources\ProspectTags\Pages;
 
-use AdvisingApp\Prospect\Filament\Resources\ProspectTagResource\Pages\CreateProspectTag;
-use AdvisingApp\Prospect\Filament\Resources\ProspectTagResource\Pages\EditProspectTag;
-use AdvisingApp\Prospect\Filament\Resources\ProspectTagResource\Pages\ListProspectTags;
-use AdvisingApp\Prospect\Filament\Resources\ProspectTagResource\Pages\ViewProspectTag;
-use AdvisingApp\Prospect\Models\Prospect;
+use AdvisingApp\Prospect\Filament\Resources\ProspectTags\ProspectTagResource;
 use App\Enums\TagType;
-use App\Filament\Clusters\ConstituentManagement;
-use App\Models\Tag;
-use Filament\Resources\Resource;
-use Illuminate\Database\Eloquent\Builder;
-use UnitEnum;
+use Filament\Forms\Components\TextInput;
+use Filament\Resources\Pages\CreateRecord;
+use Filament\Schemas\Schema;
 
-class ProspectTagResource extends Resource
+class CreateProspectTag extends CreateRecord
 {
-    protected static ?string $model = Tag::class;
+    protected static string $resource = ProspectTagResource::class;
 
-    protected static ?string $navigationLabel = 'Tags';
-
-    protected static ?int $navigationSort = 50;
-
-    protected static ?string $cluster = ConstituentManagement::class;
-
-    protected static string | UnitEnum | null $navigationGroup = 'Prospects';
-
-    public static function canAccess(): bool
+    public function form(Schema $schema): Schema
     {
-        /** @var User $user */
-        $user = auth()->user();
-
-        return $user->hasLicense(Prospect::getLicenseType());
+        return $schema
+            ->components([
+                TextInput::make('name')
+                    ->label('Name')
+                    ->required()
+                    ->maxLength(255)
+                    ->string(),
+            ]);
     }
 
-    public static function getPages(): array
+    protected function mutateFormDataBeforeCreate(array $data): array
     {
-        return [
-            'index' => ListProspectTags::route('/'),
-            'create' => CreateProspectTag::route('/create'),
-            'view' => ViewProspectTag::route('/{record}'),
-            'edit' => EditProspectTag::route('/{record}/edit'),
-        ];
-    }
+        $data['type'] = TagType::Prospect;
 
-    public static function getEloquentQuery(): Builder
-    {
-        return parent::getEloquentQuery()
-            ->where('type', TagType::Prospect);
+        return $data;
     }
 }
