+++ app-modules/prospect/src/Filament/Resources/ProspectTags/Pages/EditProspectTag.php
@@ -34,9 +34,9 @@
 </COPYRIGHT>
 */
 
-namespace AdvisingApp\Prospect\Filament\Resources\ProspectTagResource\Pages;
+namespace AdvisingApp\Prospect\Filament\Resources\ProspectTags\Pages;
 
-use AdvisingApp\Prospect\Filament\Resources\ProspectTagResource;
+use AdvisingApp\Prospect\Filament\Resources\ProspectTags\ProspectTagResource;
 use App\Filament\Resources\Pages\EditRecord\Concerns\EditPageRedirection;
 use Filament\Actions\DeleteAction;
 use Filament\Forms\Components\TextInput;
