+++ app-modules/prospect/src/Filament/Resources/ProspectTags/Pages/ViewProspectTag.php
@@ -34,9 +34,9 @@
 </COPYRIGHT>
 */
 
-namespace AdvisingApp\Prospect\Filament\Resources\ProspectTagResource\Pages;
+namespace AdvisingApp\Prospect\Filament\Resources\ProspectTags\Pages;
 
-use AdvisingApp\Prospect\Filament\Resources\ProspectTagResource;
+use AdvisingApp\Prospect\Filament\Resources\ProspectTags\ProspectTagResource;
 use Filament\Actions\EditAction;
 use Filament\Infolists\Components\TextEntry;
 use Filament\Resources\Pages\ViewRecord;
